requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
pillow
requests
numpy
torch
transformers
pandas